Create Stage 1: Card Anatomy
You should allow customers to create a new account in our banking system.

Once the program starts, you should print the menu:

1. Create an account
2. Log into account
0. Exit

If the customer chooses ‘Create an account’, you should generate a new card number which satisfies all the conditions described below.

In our banking system, credit cards should begin with 4.
The first six digits are the Issuer Identification Number (IIN). In our banking system, the IIN must be 400000.
The seventh digit to the second-to-last digit is the customer account number.
In our banking system, the customer account number can be any number, but it should be unique.
The very last digit of a credit card is the check digit or checksum. For now, the checksum can be any digit you like.

The credit card number must be unique and have a length of 16 digits.

Then you should generate a PIN code that belongs to the generated card number. A PIN code is a sequence of any 4 digits. PIN should be generated in a range from 0000 to 9999.

If the customer chooses ‘Log into account’, you should ask them to enter their card information. Your program should store all generated data until it is terminated so that a user is able to log into any of the created accounts by a card number and its pin. You can use an array to store the information.

After all information is entered correctly, you should allow the user to check the account balance; right after creating the account, the balance should be 0. It should also be possible to log out of the account and exit the program.